+++ tests/mortar/rotate_mapping.cc
@@ -97,8 +97,12 @@ test()
   dof_handler.distribute_dofs(fe);
 
   // Number of subdivisions and rotor radius
-  const auto [n_subdivisions, radius] =
-    compute_n_subdivisions_and_radius(triangulation, mortar_parameters);
+  const auto [n_subdivisions, radius, prerotation] =
+    compute_n_subdivisions_and_radius(triangulation,
+                                      mapping,
+                                      mortar_parameters);
+
+  Assert(prerotation == 0.0, ExcInternalError());
 
   // Rotate mapping
   LetheGridTools::rotate_mapping(
@@ -112,6 +116,18 @@ test()
               << std::endl;
       deallog << "Radius : " << radius << std::endl;
     }
+  // Rotate mapping
+  LetheGridTools::rotate_mapping(
+    dof_handler, mapping_cache, mapping, radius * 10, rotation_angle);
+
+  const auto [n_subdivisions1, radius1, prerotation1] =
+    compute_n_subdivisions_and_radius(triangulation,
+                                      mapping_cache,
+                                      mortar_parameters);
+
+  AssertDimension(n_subdivisions, n_subdivisions1);
+  AssertDimension(radius, radius);
+  Assert(std::abs(rotation_angle - prerotation1) < 1.e-8, ExcInternalError());
 }
 
 int